statistics: use JsonOutput for converting JesqueJobStatistic to json
If a worker can not be stopped in time during app shutdown, the worker thread eventually
gets killed in a hard way. This results in a failed job and our statistic handler gets called.
Depending on the state of the application context, this may lead to the following exception:
java.lang.RuntimeException: org.grails.web.converters.exceptions.ConverterException: Unconvertable Object of class: java.util.LinkedHashMap
	at org.grails.web.converters.AbstractConverter.toString(AbstractConverter.java:111)
	at org.grails.web.converters.AbstractConverter$toString.call(Unknown Source)
	at grails.plugins.jesque.admin.JesqueJobStatistic.asJsonString(JesqueJobStatistic.groovy:18)

So basically the grails json marshallers have been "unloaded" by the shutdown sequence already.

The fix is to use groovy's `JsonOutput` instead of grails' `as JSON`.